+++ bindings/docstrings/occupancydoc.h
